+++ src/Identity/test/Identity.Test/Passkeys/AttestedCredentialDataArgs.cs
@@ -1,0 +1,11 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+namespace Microsoft.AspNetCore.Identity.Test;
+
+internal readonly struct AttestedCredentialDataArgs
+{
+    public required ReadOnlyMemory<byte> Aaguid { get; init; }
+    public required ReadOnlyMemory<byte> CredentialId { get; init; }
+    public required ReadOnlyMemory<byte> CredentialPublicKey { get; init; }
+}
